refactor: update geo "spec" and split geo ops in ibis compiler (#2208)
