+++ bigframes/core/compile/ibis_compiler/__init__.py
@@ -21,4 +21,5 @@ registered.
 from __future__ import annotations
 
 import bigframes.core.compile.ibis_compiler.operations.generic_ops  # noqa: F401
+import bigframes.core.compile.ibis_compiler.operations.geo_ops  # noqa: F401
 import bigframes.core.compile.ibis_compiler.scalar_op_registry  # noqa: F401
